move utility functions to a new utils.ts file
